HV:transfer vmid's type to uint16_t
vmid's type prefered to be uint16_t.

V1->V2:vmid from uint32_t transfer to uint16_t
V2->V3:add range check to hcall's vmid parameter
V3->V4:seperate the declaration and actual code.
V4->V5:remove range check from hcall's vmid parameter

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/include/arch/x86/guest/vioapic.h
@@ -55,7 +55,7 @@ int	vioapic_mmio_access_handler(struct vcpu *vcpu, struct mem_io *mmio,
 		void *handler_private_data);
 
 #ifdef HV_DEBUG
-void get_vioapic_info(char *str, int str_max, int vmid);
+void get_vioapic_info(char *str, int str_max, uint16_t vmid);
 #endif /* HV_DEBUG */
 
 #endif
